+++ dag3-oplossing/assets/mole-whacked.png
