--- dag3/01-the-field/script.js
@@ -1,1 +1,0 @@
-console.log("##### Wack a mole #####");
