--- dag3/01-the-field/styles.css
@@ -1,35 +1,0 @@
-/** Don't edit **/
-*, *::before, *::after {
-  box-sizing: border-box;
-  user-select: none;
-}
-
-html, body {
-  margin: 0;
-  width: 100%;
-  height: 100%;
-  position: relative;
-  color: white;
-  display:grid;
-  place-items: center;
-  background-color: rgb(29, 105, 29);
-  font-family: 'Courier New', Courier, monospace;
-  overflow: hidden;
-}
-.field {
-  width: 600px;
-  height: 600px;
-  background-color: rgb(158, 92, 27);
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  gap: 60px;
-  padding: 20px;
-}
-/** END Don't edit */
-
-
-/* Excersice 1: Let's draw the holes */
-.hole {
-  /* TODO: Add styles */
-}
