--- dag3/02-interactive-hammer/styles.css
@@ -1,45 +1,0 @@
-/** Don't edit **/
-*, *::before, *::after {
-  box-sizing: border-box;
-  user-select: none;
-}
-
-html, body {
-  margin: 0;
-  width: 100%;
-  height: 100%;
-  position: relative;
-  color: white;
-  display:grid;
-  place-items: center;
-  background-color: rgb(29, 105, 29);
-  font-family: 'Courier New', Courier, monospace;
-  overflow: hidden;
-}
-.field {
-  width: 600px;
-  height: 600px;
-  background-color: rgb(158, 92, 27);
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  gap: 60px;
-  padding: 20px;
-}
-/** END Don't edit */
-
-.hole {
-  border: 1px solid rgb(50, 46, 43);
-  border-radius: 100%;
-  background: rgb(82, 66, 50);
-  box-shadow: inset 0 0 10px 0 rgb(70, 43, 16);
-}
-
-/* Excersice 2: Make the hammer interactive */
-.hammer {
-
-}
-
-.hammer.active {
-  
-}
