--- dag3/03-the-mole/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Wack-A-Mole | Codeo Dag 3</title>
-
-    <link href="./styles.css" rel="stylesheet">
-  </head>
-  <body>
-      <div class="world">
-        <H1>Wack a mole</H1>
-        
-        <div class="field">
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-        </div>  
-      </div>
-      <div class="hammer"></div>
-      <script src="./script.js"></script>
-  </body>
-</html>
