--- dag3/04-timeouts/script.js
@@ -1,41 +1,0 @@
-console.log("##### Wack a mole #####");
-
-// We have two selectors we can use to find our elements on a page
-// querySelector will return the first element it finds
-// querySelectorAll will return all the elements it finds
-let holes = document.querySelectorAll(".hole");
-let hammer = document.querySelector(".hammer");
-
-function run() {
-  // Get a random hole
-  let hole = holes[Math.floor(Math.random() * holes.length)];
-
-  // Add the mole to that random hole
-  let img = document.createElement("img");
-  img.classList.add("mole");
-  img.src = "assets/mole.png";
-
-  img.addEventListener("click", () => {
-    console.log("Whack!");
-    hole.removeChild(img);
-    run();
-  });
-
-  hole.appendChild(img);
-}
-
-run();
-
-// Adding events
-window.addEventListener("mousemove", (event) => {
-  hammer.style.left = event.pageX + "px";
-  hammer.style.top = event.pageY + "px";
-});
-
-window.addEventListener("mousedown", (event) => {
-  hammer.classList.add("active");
-});
-
-window.addEventListener("mouseup", (event) => {
-  hammer.classList.remove("active");
-});
