--- dag3/06-gameloop/index.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Wack-A-Mole | Codeo Dag 3</title>
-
-    <link href="./styles.css" rel="stylesheet">
-  </head>
-  <body>
-      <div class="world">
-        <H1>Wack a mole</H1>
-        <p>
-          <span class="score">Score: <span>0</span>
-        </p>
-        <div class="field">
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-          <div class="hole"></div>
-        </div>  
-      </div>
-      <div class="hammer"></div>
-      <script src="./script.js"></script>
-  </body>
-</html>
