+++ dag3/README.md
@@ -1,15 +1,10 @@
 # Codeo Dag 3
 
 Zorg dat je de instructies van dag 2 gevolgd hebt, en NodeJS geïnstalleerd hebt
-op je computer. Hierna kun je alle 6 de stukjes code één voor één draaien met de 
-volgende commando's in de terminal:
+op je computer. Hierna kan je de code voor vandaag draaien door het volgende in een terminal te typen:
 
-npm run start-1
-npm run start-2
-npm run start-3
-npm run start-4
-npm run start-5
-npm run start-6
+```shell
+npm run start
+```
 
-Kies welk stukje je wil draaien, waarschijnlijk heb je ze niet alle 6 tegelijk 
-nodig.
+Hierna wordt Wack-A-Mole automatisch geopend in je browser. Je kan jouw eigen code toevoegen in de `index.html`, `main.js` en `styles.css` bestanden.
