+++ dag3/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Wack-A-Mole | Codeo Dag 3</title>
+
+    <link href="base.css" rel="stylesheet">
+    <link href="styles.css" rel="stylesheet">
+</head>
+<body>
+<div class="world">
+    <H1>Wack-A-Mole</H1>
+    <p>
+        <span class="score">Score: <span>0</span></span>
+    </p>
+    <div class="field">
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+        <div class="hole"></div>
+    </div>
+</div>
+<div class="hammer"></div>
+
+<script src="main.js" type="module"></script>
+</body>
+</html>
