+++ dag3/main.js
@@ -1,0 +1,7 @@
+import getRandomHole from "./random-hole.js";
+
+console.log("##### Wack a mole #####");
+
+/**
+ * You can get a random hole using the getRandomHole() function, by calling it with the list of hole elements.
+ */
