+++ dag3/random-hole.js
@@ -1,0 +1,5 @@
+const getRandomHole = (holes) => {
+    return holes[Math.floor(Math.random() * holes.length)];
+};
+
+export default getRandomHole;
