+++ dag3/styles.css
@@ -1,0 +1,17 @@
+.hole {
+
+}
+
+.hammer {
+
+}
+
+.hammer.active {
+
+}
+
+/* Exercise 3: Style the mole */
+.mole {
+
+}
+
